feat!: Simplified API to perform transactions
The previous version represented the concept of
a transaction with the Transaction class, which has now
been removed. This version keeps the [withTransaction] method,
but the callback receives now a session instead of a transaction.
The Session.startTransaction now does not return an instance
of transactions, but it now has the commitTransaction and
abortTransaction methods.

Besides being a simpler API, the internal implementation
became a whole lot simpler, making it more reliable.